+++ svelte.config.js
@@ -24,7 +24,7 @@ const config = {
       $lib: 'src/lib'
     },
     serviceWorker: {
-      register: false
+      register: true
     }
   }
 };